working

+++ src/main/java/com/gcash/courier/delivery/model/Parcel.java
@@ -3,10 +3,10 @@ package com.gcash.courier.delivery.model;
 
 public record Parcel(
 
-        Double weight,
-        Double height,
-        Double width,
-        Double length,
+        float weight,
+        float height,
+        float width,
+        float length,
         String voucherCode
 ) {
 }
